# Slotted array vectorized