--- a/src/Paprika.Tests/Data/SlottedArrayTests.cs
+++ b/src/Paprika.Tests/Data/SlottedArrayTests.cs
@@ -1,8 +1,6 @@
 using FluentAssertions;
-using NUnit.Framework;
 using Paprika.Crypto;
 using Paprika.Data;
-using Paprika.Merkle;
 
 namespace Paprika.Tests.Data;