--- a/src/Paprika.Tests/Data/SlottedArrayTests.cs
+++ b/src/Paprika.Tests/Data/SlottedArrayTests.cs
@@ -1,6 +1,9 @@
+using System.Runtime.InteropServices;
+using System.Runtime.Intrinsics;
 using FluentAssertions;
 using Paprika.Crypto;
 using Paprika.Data;
+using Paprika.Store;
 
 namespace Paprika.Tests.Data;
 
@@ -20,7 +23,7 @@ public class SlottedArrayTests
     {
         var key0 = Values.Key0.Span;
 
-        Span<byte> span = stackalloc byte[48];
+        Span<byte> span = stackalloc byte[SlottedArray.MinimalSizeWithNoData + key0.Length + Data0.Length];
         var map = new SlottedArray(span);
 
         map.SetAssert(key0, Data0);
@@ -118,7 +121,7 @@ public class SlottedArrayTests
     {
         var key0 = Values.Key0.Span;
 
-        Span<byte> span = stackalloc byte[48];
+        Span<byte> span = stackalloc byte[128];
         var map = new SlottedArray(span);
 
         var data = ReadOnlySpan<byte>.Empty;
@@ -149,7 +152,7 @@ public class SlottedArrayTests
     public Task Defragment_when_no_more_space()
     {
         // by trial and error, found the smallest value that will allow to put these two
-        Span<byte> span = stackalloc byte[88];
+        Span<byte> span = stackalloc byte[SlottedArray.MinimalSizeWithNoData + 88];
         var map = new SlottedArray(span);
 
         var key0 = Values.Key0.Span;
@@ -177,7 +180,7 @@ public class SlottedArrayTests
     public Task Update_in_situ()
     {
         // by trial and error, found the smallest value that will allow to put these two
-        Span<byte> span = stackalloc byte[48];
+        Span<byte> span = stackalloc byte[128];
         var map = new SlottedArray(span);
 
         var key1 = Values.Key1.Span;
@@ -194,11 +197,11 @@ public class SlottedArrayTests
     [Test]
     public Task Update_in_resize()
     {
-        // Update the value, with the next one being bigger.
-        Span<byte> span = stackalloc byte[56];
-        var map = new SlottedArray(span);
-
         var key0 = Values.Key0.Span;
+
+        // Update the value, with the next one being bigger.
+        Span<byte> span = stackalloc byte[SlottedArray.MinimalSizeWithNoData + key0.Length + Data0.Length];
+        var map = new SlottedArray(span);
 
         map.SetAssert(key0, Data0);
         map.SetAssert(key0, Data2);
@@ -212,7 +215,7 @@ public class SlottedArrayTests
     [Test]
     public Task Small_keys_compression()
     {
-        Span<byte> span = stackalloc byte[256];
+        Span<byte> span = stackalloc byte[512];
         var map = new SlottedArray(span);
 
         Span<byte> key = stackalloc byte[1];
@@ -240,6 +243,81 @@ public class SlottedArrayTests
 
         // verify
         return Verify(span.ToArray());
+    }
+
+    [Test(Description = "Make a lot of requests to make breach the vector count")]
+    public void Breach_VectorSize_with_key_count()
+    {
+        const int seed = 13;
+        var random = new Random(seed);
+        Span<byte> key = stackalloc byte[4];
+
+        var map = new SlottedArray(new byte[Page.PageSize]);
+
+        const int count = 257;
+
+        for (var i = 0; i < count; i++)
+        {
+            random.NextBytes(key);
+            map.SetAssert(key, [(byte)(i & 255)]);
+        }
+
+        // reset
+        random = new Random(seed);
+        for (var i = 0; i < count; i++)
+        {
+            random.NextBytes(key);
+            map.GetAssert(key, [(byte)(i & 255)]);
+        }
+    }
+
+    [Test]
+    public void Roll_over()
+    {
+        const int seed = 13;
+        var random = new Random(seed);
+        Span<byte> key = stackalloc byte[4];
+
+        var map = new SlottedArray(new byte[1024]);
+
+        byte count = 0;
+
+        random.NextBytes(key);
+        while (map.TrySet(NibblePath.FromKey(key), [count]))
+        {
+            count++;
+            random.NextBytes(key);
+        }
+
+        // reset, delete some
+        //random = new Random(seed);
+
+        using var e = map.EnumerateAll();
+        for (var i = 0; i < count; i++)
+        {
+            //random.NextBytes(key);
+            e.MoveNext().Should().BeTrue();
+
+            if (ShouldBeDeleted(i))
+            {
+                map.Delete(e.Current);
+                //map.Delete(NibblePath.FromKey(key)).Should().BeTrue();
+            }
+        }
+
+        // reset, assert
+        random = new Random(seed);
+        for (var i = 0; i < count; i++)
+        {
+            random.NextBytes(key);
+
+            var exist = map.TryGet(NibblePath.FromKey(key), out var data);
+            exist.Should().NotBe(ShouldBeDeleted(i));
+        }
+
+        return;
+
+        static bool ShouldBeDeleted(int i) => i % 2 == 0;
     }
 
     private static ReadOnlySpan<byte> Data(byte key) => new[] { key };
@@ -452,14 +530,14 @@ public class SlottedArrayTests
     public void Move_to_8()
     {
         var original = new SlottedArray(stackalloc byte[512]);
-        var copy0 = new SlottedArray(stackalloc byte[64]);
-        var copy1 = new SlottedArray(stackalloc byte[64]);
-        var copy2 = new SlottedArray(stackalloc byte[64]);
-        var copy3 = new SlottedArray(stackalloc byte[64]);
-        var copy4 = new SlottedArray(stackalloc byte[64]);
-        var copy5 = new SlottedArray(stackalloc byte[64]);
-        var copy6 = new SlottedArray(stackalloc byte[64]);
-        var copy7 = new SlottedArray(stackalloc byte[64]);
+        var copy0 = new SlottedArray(stackalloc byte[128]);
+        var copy1 = new SlottedArray(stackalloc byte[128]);
+        var copy2 = new SlottedArray(stackalloc byte[128]);
+        var copy3 = new SlottedArray(stackalloc byte[128]);
+        var copy4 = new SlottedArray(stackalloc byte[128]);
+        var copy5 = new SlottedArray(stackalloc byte[128]);
+        var copy6 = new SlottedArray(stackalloc byte[128]);
+        var copy7 = new SlottedArray(stackalloc byte[128]);
 
         var key0 = NibblePath.Empty;
         var key1 = NibblePath.Parse("1");