… docs/design.md
… src/Paprika.Tests/Data/SlottedArrayTests.Defragment_when_no_more_space.verified.bin
… src/Paprika.Tests/Data/SlottedArrayTests.Enumerate_all_odd=0.verified.bin
… src/Paprika.Tests/Data/SlottedArrayTests.Enumerate_all_odd=1.verified.bin
… src/Paprika.Tests/Data/SlottedArrayTests.Enumerate_long_key_oddStart=0_lengthCutOff=0.verified.bin
… src/Paprika.Tests/Data/SlottedArrayTests.Enumerate_long_key_oddStart=0_lengthCutOff=1.verified.bin
… src/Paprika.Tests/Data/SlottedArrayTests.Enumerate_long_key_oddStart=1_lengthCutOff=0.verified.bin
… src/Paprika.Tests/Data/SlottedArrayTests.Enumerate_long_key_oddStart=1_lengthCutOff=1.verified.bin
… src/Paprika.Tests/Data/SlottedArrayTests.Set_Get_Delete_Get_AnotherSet.verified.bin
… src/Paprika.Tests/Data/SlottedArrayTests.Set_Get_Empty.verified.bin
… src/Paprika.Tests/Data/SlottedArrayTests.Small_keys_compression.verified.bin
… src/Paprika.Tests/Data/SlottedArrayTests.Update_in_resize.verified.bin
… src/Paprika.Tests/Data/SlottedArrayTests.Update_in_situ.verified.bin
… src/Paprika.Tests/Store/DbTests.cs
… src/Paprika/Data/SlottedArray.cs